requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Dance==7.1.0
psycopg2-binary==2.9.10
SQLAlchemy==2.0.41
Werkzeug==3.1.3
email-validator==2.2.0
oauthlib==3.3.1
PyJWT==2.10.1
openai==1.93.0
gunicorn==23.0.0
Flask-Migrate==4.0.5
python-dotenv==1.0.0
requests==2.31.0
flask-socketio
eventlet 